requirements.txt
baidu-aip==2.2.5.2